fix(quota): correct budget calculation, diagnostics attribution, and Python 3.13 compatibility
- Fix calculate_remaining_polling_budget: use actual_polls_today as baseline for
  inferred_external detection, eliminating double-penalty on external calls
- Fix calculate_safety_reserve_interval: distribute over 3-hour safe window
  (was 1 hour, causing 3x too-frequent polling during reset window)
- Fix available_now to use effective_threshold instead of throttle_threshold
- Fix diagnostics user_so_far/user_excess: zone polls were incorrectly attributed
  as external user calls; now uses tracked poll counter
- Fix _update_learned_window dead code: single_observation branch was unreachable
- Track actual polling calls in _polling_calls_today per fetch, reset on quota reset
- Align tooling with Python 3.13: CI, ruff target-version, pyproject.toml

+++ custom_components/tado_hijack/number.py
@@ -24,6 +24,7 @@ from .models import TadoEntityDefinition
 if TYPE_CHECKING:
     from homeassistant.config_entries import ConfigEntry
     from homeassistant.core import HomeAssistant
+
     from .coordinator import TadoDataUpdateCoordinator
 
 
@@ -69,7 +70,7 @@ class TadoOptimisticNumber(TadoOptimisticMixin, RestoreEntity, NumberEntity):
             val is not None
             and getattr(self, "_attr_suggested_display_precision", None) == 0
         ):
-            return int(round(float(val)))
+            return round(float(val))
 
         return float(val) if val is not None else None
